Exit if bootstrap log contains errors
Related-Bug: #1317513

+++ iso/bootstrap_admin_node.docker.sh
@@ -77,6 +77,11 @@ cp -a /etc/astute.yaml /etc/fuel/astute.yaml
 # apply puppet
 # LANG variable is a workaround for puppet-3.4.2 bug. See LP#1312758 for details
 puppet apply -d -v /etc/puppet/modules/nailgun/examples/host-only.pp
+if grep -q Error /var/log/puppet/bootstrap_admin_node.log
+then
+    echo "Something went wrong during puppet run..."
+    exit 255
+fi
 rmdir /var/log/remote && ln -s /var/log/docker-logs/remote /var/log/remote
 
 echo -n "Waiting for Fuel UI to be ready..."
